Pacman Clone made in early days
Yes, this clone was made while I was learning Javascript. So the code is
very messy and not made with normal Javascript conventions. I'm planning
on fixing the code up a bit to make it more "presentable", but I still
think it's something cool to show. Even if the code is [very] messy, it
did it's job really well and the result is something I'm happy with.